class diagram update, jwtpassphrase move in new .env.local, errors responses move in a service and appcontroller split in two controllers
