+++ .env
@@ -36,5 +36,5 @@ DATABASE_URL="mysql://root:@127.0.0.1:3306/bilemo?serverVersion=5.7"
 ###> lexik/jwt-authentication-bundle ###
 JWT_SECRET_KEY=%kernel.project_dir%/config/jwt/private.pem
 JWT_PUBLIC_KEY=%kernel.project_dir%/config/jwt/public.pem
-JWT_PASSPHRASE=happyapi
+JWT_PASSPHRASE=
 ###< lexik/jwt-authentication-bundle ###
